Share common functions initPython, addToPythonPath and reportPythonError between PluginManager and PythonEngine
