+++ src/EnergyPlus/PluginManager.cc
@@ -358,92 +358,6 @@ void PluginManager::setupOutputVariables([[maybe_unused]] EnergyPlusData &state)
 } // setupOutputVariables()
 
 #if LINK_WITH_PYTHON
-void initPython(EnergyPlusData &state, fs::path const &pathToPythonPackages)
-{
-    PyStatus status;
-
-    // first pre-config Python so that it can speak UTF-8
-    PyPreConfig preConfig;
-    // This is the other related line that caused Decent CI to start having trouble.  I'm putting it back to
-    // PyPreConfig_InitPythonConfig, even though I think it should be isolated.  Will deal with this after IO freeze.
-    PyPreConfig_InitPythonConfig(&preConfig);
-    // PyPreConfig_InitIsolatedConfig(&preConfig);
-    preConfig.utf8_mode = 1;
-    status = Py_PreInitialize(&preConfig);
-    if (PyStatus_Exception(status) != 0) {
-        ShowFatalError(state, std::format("Could not pre-initialize Python to speak UTF-8... {}", status));
-    }
-
-    PyConfig config;
-    PyConfig_InitIsolatedConfig(&config);
-    config.isolated = 1;
-
-    status = PyConfig_SetBytesString(&config, &config.program_name, PluginManagement::programName);
-    if (PyStatus_Exception(status) != 0) {
-        ShowFatalError(state, std::format("Could not initialize program_name on PyConfig... {}", status));
-    }
-
-    status = PyConfig_Read(&config);
-    if (PyStatus_Exception(status) != 0) {
-        ShowFatalError(state, std::format("Could not read back the PyConfig... {}", status));
-    }
-
-    // ReSharper disable once CppRedundantTypenameKeyword
-    if constexpr (std::is_same_v<typename fs::path::value_type, wchar_t>) {
-        // PyConfig_SetString copies the wide character string str into *config_str.
-        // ReSharper disable once CppDFAUnreachableCode
-        std::wstring const ws = pathToPythonPackages.generic_wstring();
-        const wchar_t *wcharPath = ws.c_str();
-
-        status = PyConfig_SetString(&config, &config.home, wcharPath);
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not set home to {:g} on PyConfig... {}", pathToPythonPackages, status));
-        }
-        status = PyConfig_SetString(&config, &config.base_prefix, wcharPath);
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not set base_prefix to {:g} on PyConfig... {}", pathToPythonPackages, status));
-        }
-        config.module_search_paths_set = 1;
-        status = PyWideStringList_Append(&config.module_search_paths, wcharPath);
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not add {:g} to module_search_paths on PyConfig... {}", pathToPythonPackages, status));
-        }
-
-    } else {
-        // PyConfig_SetBytesString takes a `const char * str` and decodes str using Py_DecodeLocale() and set the result into *config_str
-        // But we want to avoid doing it three times, so we PyDecodeLocale manually
-        // Py_DecodeLocale can be called because Python has been PreInitialized.
-        wchar_t *wcharPath = Py_DecodeLocale(pathToPythonPackages.generic_string().c_str(), nullptr); // This allocates!
-
-        status = PyConfig_SetString(&config, &config.home, wcharPath);
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not set home to {:g} on PyConfig... {}", pathToPythonPackages, status));
-        }
-        status = PyConfig_SetString(&config, &config.base_prefix, wcharPath);
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not set base_prefix to {:g} on PyConfig... {}", pathToPythonPackages, status));
-        }
-        config.module_search_paths_set = 1;
-        status = PyWideStringList_Append(&config.module_search_paths, wcharPath);
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not add {:g} to module_search_paths on PyConfig... {}", pathToPythonPackages, status));
-        }
-
-        PyMem_RawFree(wcharPath);
-    }
-
-    // This was Py_InitializeFromConfig(&config), but was giving a seg fault when running inside
-    // another Python instance, for example as part of an API run.  Per the example here:
-    // https://docs.python.org/3/c-api/init_config.html#preinitialize-python-with-pypreconfig
-    // It looks like we don't need to initialize from config again, it should be all set up with
-    // the init calls above, so just initialize and move on.
-    // UPDATE: This worked happily for me on Linux, and also when I build locally on Windows, but not on Decent CI
-    // I suspect a difference in behavior for Python versions.  I'm going to temporarily revert this back to initialize
-    // with config and get IO freeze going, then get back to solving it.
-    // Py_Initialize();
-    Py_InitializeFromConfig(&config);
-}
-
 // GilGrabber is an RAII helper that will ensure we release the GIL (including if we end up throwing)
 struct GilGrabber
 {
@@ -477,9 +391,8 @@ PluginManager::PluginManager(EnergyPlusData &state) : eplusRunningViaPythonAPI(s
     } else {
         programDir = FileSystem::getParentDirectoryPath(FileSystem::getAbsolutePath(FileSystem::getProgramPath()));
     }
-    fs::path const pathToPythonPackages = programDir / "python_lib";
 
-    initPython(state, pathToPythonPackages);
+    EnergyPlus::PythonHelpers::initPython(state, programDir);
 
     // Take control of the global interpreter lock, which will be released via RAII
     GilGrabber gil_grabber;
@@ -487,17 +400,10 @@ PluginManager::PluginManager(EnergyPlusData &state) : eplusRunningViaPythonAPI(s
     // call this once to allow us to add to, and report, sys.path later as needed
     PyRun_SimpleString("import sys"); // allows us to report sys.path later
 
-    // we also need to set an extra import path to find some dynamic library loading stuff, again make it relative to the binary
-    addToPythonPath(state, programDir / "python_lib/lib-dynload", false);
-
     // now for additional paths:
-    // we'll always want to add the program executable directory to PATH so that Python can find the installed pyenergyplus package
     // we will then optionally add the current working directory to allow Python to find scripts in the current directory
     // we will then optionally add the directory of the running IDF to allow Python to find scripts kept next to the IDF
     // we will then optionally add any additional paths the user specifies on the search paths object
-
-    // so add the executable directory here
-    addToPythonPath(state, programDir, false);
 
     // Read all the additional search paths next
     std::string const sPaths = "PythonPlugin:SearchPaths";
@@ -521,7 +427,7 @@ PluginManager::PluginManager(EnergyPlusData &state) : eplusRunningViaPythonAPI(s
                 // defaulted to YES
             }
             if (workingDirFlagUC == "YES") {
-                addToPythonPath(state, ".", false);
+                EnergyPlus::PythonHelpers::addToPythonPath(state, ".", false);
             }
             std::string inputFileDirFlagUC = "YES";
             try {
@@ -530,7 +436,7 @@ PluginManager::PluginManager(EnergyPlusData &state) : eplusRunningViaPythonAPI(s
                 // defaulted to YES
             }
             if (inputFileDirFlagUC == "YES") {
-                addToPythonPath(state, state.dataStrGlobals->inputDirPath, false);
+                EnergyPlus::PythonHelpers::addToPythonPath(state, state.dataStrGlobals->inputDirPath, false);
             }
 
             std::string epInDirFlagUC = "YES";
@@ -550,7 +456,7 @@ PluginManager::PluginManager(EnergyPlusData &state) : eplusRunningViaPythonAPI(s
                 } else {
                     fs::path const epinRootDir = FileSystem::getParentDirectoryPath(fs::path(epinPathObject));
                     if (FileSystem::pathExists(epinRootDir)) {
-                        addToPythonPath(state, epinRootDir, true);
+                        EnergyPlus::PythonHelpers::addToPythonPath(state, epinRootDir, true);
                     } else {
                         ShowWarningMessage(state,
                                            "PluginManager: Search path inputs requested adding epin variable to Python path, but epin "
@@ -563,7 +469,7 @@ PluginManager::PluginManager(EnergyPlusData &state) : eplusRunningViaPythonAPI(s
                 auto const &vars = fields.at("py_search_paths");
                 for (const auto &var : vars) {
                     try {
-                        addToPythonPath(state, fs::path(var.at("search_path").get<std::string>()), true);
+                        EnergyPlus::PythonHelpers::addToPythonPath(state, fs::path(var.at("search_path").get<std::string>()), true);
                     } catch ([[maybe_unused]] const nlohmann::json::out_of_range &e) {
                         // empty entry
                     }
@@ -575,8 +481,8 @@ PluginManager::PluginManager(EnergyPlusData &state) : eplusRunningViaPythonAPI(s
         }
     } else {
         // no search path objects in the IDF, just do the default behavior: add the current working dir and the input file dir, + epin env var
-        addToPythonPath(state, ".", false);
-        addToPythonPath(state, state.dataStrGlobals->inputDirPath, false);
+        EnergyPlus::PythonHelpers::addToPythonPath(state, ".", false);
+        EnergyPlus::PythonHelpers::addToPythonPath(state, state.dataStrGlobals->inputDirPath, false);
 
         std::string epin_path; // NOLINT(misc-const-correctness)
         get_environment_variable("epin", epin_path);
@@ -584,7 +490,7 @@ PluginManager::PluginManager(EnergyPlusData &state) : eplusRunningViaPythonAPI(s
         if (!epinPathObject.empty()) {
             fs::path const epinRootDir = FileSystem::getParentDirectoryPath(fs::path(epinPathObject));
             if (FileSystem::pathExists(epinRootDir)) {
-                addToPythonPath(state, epinRootDir, true);
+                EnergyPlus::PythonHelpers::addToPythonPath(state, epinRootDir, true);
             }
         }
     }
@@ -704,79 +610,6 @@ PluginInstance::PluginInstance(const fs::path &_modulePath, const std::string &_
 {
 }
 
-void PluginInstance::reportPythonError([[maybe_unused]] EnergyPlusData &state)
-{
-#if LINK_WITH_PYTHON
-    PyObject *exc_type = nullptr;
-    PyObject *exc_value = nullptr;
-    PyObject *exc_tb = nullptr;
-    PyErr_Fetch(&exc_type, &exc_value, &exc_tb);
-    // Normalizing the exception is needed. Without it, our custom EnergyPlusException go through just fine
-    // but any ctypes built-in exception for eg will have wrong types
-    PyErr_NormalizeException(&exc_type, &exc_value, &exc_tb);
-    PyObject *str_exc_value = PyObject_Repr(exc_value); // Now a unicode object
-    PyObject *pyStr2 = PyUnicode_AsEncodedString(str_exc_value, "utf-8", "Error ~");
-    Py_DECREF(str_exc_value);
-    char *strExcValue = PyBytes_AsString(pyStr2); // NOLINT(hicpp-signed-bitwise)
-    Py_DECREF(pyStr2);
-    ShowContinueError(state, "Python error description follows: ");
-    ShowContinueError(state, strExcValue);
-
-    // See if we can get a full traceback.
-    // Calls into python, and does the same as capturing the exception in `e`
-    // then `print(traceback.format_exception(e.type, e.value, e.tb))`
-    PyObject *pModuleName = PyUnicode_DecodeFSDefault("traceback");
-    PyObject *pyth_module = PyImport_Import(pModuleName);
-    Py_DECREF(pModuleName);
-
-    if (pyth_module == nullptr) {
-        ShowContinueError(state, "Cannot find 'traceback' module in reportPythonError(), this is weird");
-        return;
-    }
-
-    PyObject *pyth_func = PyObject_GetAttrString(pyth_module, "format_exception");
-    Py_DECREF(pyth_module); // PyImport_Import returns a new reference, decrement it
-
-    if ((pyth_func != nullptr) || (PyCallable_Check(pyth_func) != 0)) {
-
-        PyObject *pyth_val = PyObject_CallFunction(pyth_func, "OOO", exc_type, exc_value, exc_tb);
-
-        // traceback.format_exception returns a list, so iterate on that
-        if ((pyth_val == nullptr) || !PyList_Check(pyth_val)) { // NOLINT(hicpp-signed-bitwise)
-            ShowContinueError(state, "In reportPythonError(), traceback.format_exception did not return a list.");
-            return;
-        }
-
-        Py_ssize_t numVals = PyList_Size(pyth_val);
-        if (numVals == 0) {
-            ShowContinueError(state, "No traceback available");
-            return;
-        }
-
-        ShowContinueError(state, "Python traceback follows: ");
-        ShowContinueError(state, "```");
-
-        for (Py_ssize_t itemNum = 0; itemNum < numVals; itemNum++) {
-            PyObject *item = PyList_GetItem(pyth_val, itemNum);
-            if (PyUnicode_Check(item)) { // NOLINT(hicpp-signed-bitwise) -- something inside Python code causes warning
-                std::string traceback_line = PyUnicode_AsUTF8(item);
-                if (!traceback_line.empty() && traceback_line[traceback_line.length() - 1] == '\n') {
-                    traceback_line.erase(traceback_line.length() - 1);
-                }
-                ShowContinueError(state, std::format(" >>> {}", traceback_line));
-            }
-            // PyList_GetItem returns a borrowed reference, do not decrement
-        }
-
-        ShowContinueError(state, "```");
-
-        // PyList_Size returns a borrowed reference, do not decrement
-        Py_DECREF(pyth_val); // PyObject_CallFunction returns new reference, decrement
-    }
-    Py_DECREF(pyth_func); // PyObject_GetAttrString returns a new reference, decrement it
-#endif
-}
-
 void PluginInstance::setup([[maybe_unused]] EnergyPlusData &state)
 {
 #if LINK_WITH_PYTHON
@@ -804,7 +637,7 @@ void PluginInstance::setup([[maybe_unused]] EnergyPlusData &state)
         ShowContinueError(state, std::format("Current sys.path={}", PluginManager::currentPythonPath()));
         // ONLY call PyErr_Print if PyErr has occurred, otherwise it will cause other problems
         if (PyErr_Occurred() != nullptr) {
-            reportPythonError(state);
+            EnergyPlus::PythonHelpers::reportPythonError(state);
         } else {
             ShowContinueError(state, "It could be that the module could not be found, or that there was an error in importing");
         }
@@ -814,7 +647,7 @@ void PluginInstance::setup([[maybe_unused]] EnergyPlusData &state)
     if (pModuleDict == nullptr) {
         ShowSevereError(state, std::format("Failed to read module dictionary from module \"{:g}\"", this->modulePath));
         if (PyErr_Occurred() != nullptr) {
-            reportPythonError(state);
+            EnergyPlus::PythonHelpers::reportPythonError(state);
         } else {
             ShowContinueError(state, "It could be that the module was empty");
         }
@@ -836,7 +669,7 @@ void PluginInstance::setup([[maybe_unused]] EnergyPlusData &state)
     if (pClass == nullptr) {
         ShowSevereError(state, std::format(R"(Failed to get class type "{}" from module "{:g}")", className, modulePath));
         if (PyErr_Occurred() != nullptr) {
-            reportPythonError(state);
+            EnergyPlus::PythonHelpers::reportPythonError(state);
         } else {
             ShowContinueError(state, "It could be the class name is misspelled or missing.");
         }
@@ -845,7 +678,7 @@ void PluginInstance::setup([[maybe_unused]] EnergyPlusData &state)
     if (PyCallable_Check(pClass) == 0) {
         ShowSevereError(state, std::format("Got class type \"{}\", but it cannot be called/instantiated", className));
         if (PyErr_Occurred() != nullptr) {
-            reportPythonError(state);
+            EnergyPlus::PythonHelpers::reportPythonError(state);
         } else {
             ShowContinueError(state, "Is it possible the class name is actually just a variable?");
         }
@@ -856,7 +689,7 @@ void PluginInstance::setup([[maybe_unused]] EnergyPlusData &state)
     if (this->pClassInstance == nullptr) {
         ShowSevereError(state, std::format("Something went awry calling class constructor for class \"{}\"", className));
         if (PyErr_Occurred() != nullptr) {
-            reportPythonError(state);
+            EnergyPlus::PythonHelpers::reportPythonError(state);
         } else {
             ShowContinueError(state, "It is possible the plugin class constructor takes extra arguments - it shouldn't.");
         }
@@ -876,7 +709,7 @@ void PluginInstance::setup([[maybe_unused]] EnergyPlusData &state)
             std::format(
                 R"(Could not find or call function "{}" on class "{:g}.{}")", detectOverriddenFunctionName, this->modulePath, this->className));
         if (PyErr_Occurred() != nullptr) {
-            reportPythonError(state);
+            EnergyPlus::PythonHelpers::reportPythonError(state);
         } else {
             ShowContinueError(state, "This function should be available on the base class, so this is strange.");
         }
@@ -887,7 +720,7 @@ void PluginInstance::setup([[maybe_unused]] EnergyPlusData &state)
     if (pFunctionResponse == nullptr) {
         ShowSevereError(state, std::format("Call to _detect_overridden() on {} failed!", this->stringIdentifier));
         if (PyErr_Occurred() != nullptr) {
-            reportPythonError(state);
+            EnergyPlus::PythonHelpers::reportPythonError(state);
         } else {
             ShowContinueError(state, "This is available on the base class and should not be overridden...strange.");
         }
@@ -1147,7 +980,7 @@ bool PluginInstance::run(EnergyPlusData &state, EMSManager::EMSCallFrom iCalledF
         std::string const functionNameAsString(functionName); // only convert to string if an error occurs
         ShowSevereError(state, std::format("Call to {}() on {} failed!", functionNameAsString, this->stringIdentifier));
         if (PyErr_Occurred() != nullptr) {
-            reportPythonError(state);
+            EnergyPlus::PythonHelpers::reportPythonError(state);
         } else {
             ShowContinueError(state, "This could happen for any number of reasons, check the plugin code.");
         }
@@ -1193,56 +1026,10 @@ std::vector<std::string> PluginManager::currentPythonPath()
     }
     return pathLibs;
 }
-
-void PluginManager::addToPythonPath(EnergyPlusData &state, const fs::path &includePath, bool userDefinedPath)
-{
-    if (includePath.empty()) {
-        return;
-    }
-
-    // We use generic_string / generic_wstring here, which will always use a forward slash as directory separator even on windows
-    // This doesn't handle the (very strange, IMHO) case were on unix you have backlashes (which are VALID filenames on Unix!)
-    // Could use FileSystem::makeNativePath first to convert the backslashes to forward slashes on Unix
-    PyObject *unicodeIncludePath;
-    // ReSharper disable once CppRedundantTypenameKeyword
-    if constexpr (std::is_same_v<typename fs::path::value_type, wchar_t>) {
-        // ReSharper disable once CppDFAUnreachableCode
-        const std::wstring ws = includePath.generic_wstring();
-        unicodeIncludePath = PyUnicode_FromWideChar(ws.c_str(), static_cast<Py_ssize_t>(ws.size())); // New reference
-    } else {
-        const std::string s = includePath.generic_string();
-        unicodeIncludePath = PyUnicode_FromString(s.c_str()); // New reference
-    }
-    if (unicodeIncludePath == nullptr) {
-        ShowFatalError(state, std::format("ERROR converting the path \"{:g}\" for addition to the sys.path in Python", includePath));
-    }
-
-    PyObject *sysPath = PySys_GetObject("path"); // Borrowed reference
-    int const ret = PyList_Insert(sysPath, 0, unicodeIncludePath);
-    Py_DECREF(unicodeIncludePath);
-
-    if (ret != 0) {
-        if (PyErr_Occurred() != nullptr) {
-            PluginInstance::reportPythonError(state);
-        }
-        ShowFatalError(state, std::format("ERROR adding \"{:g}\" to the sys.path in Python", includePath));
-    }
-
-    if (userDefinedPath) {
-        ShowMessage(state, std::format("Successfully added path \"{:g}\" to the sys.path in Python", includePath));
-    }
-
-    // PyRun_SimpleString)("print(' EPS : ' + str(sys.path))");
-}
 #else
 std::vector<std::string> PluginManager::currentPythonPath()
 {
     return {};
-}
-void PluginManager::addToPythonPath([[maybe_unused]] EnergyPlusData &state,
-                                    [[maybe_unused]] const fs::path &path,
-                                    [[maybe_unused]] bool userDefinedPath)
-{
 }
 #endif
