+++ src/EnergyPlus/PluginManager.hh
@@ -70,8 +70,6 @@ struct EnergyPlusData;
 
 namespace PluginManagement {
 
-    constexpr const char *programName = "python";
-
     void registerNewCallback(const EnergyPlusData &state, EMSManager::EMSCallFrom iCalledFrom, const std::function<void(void *)> &f);
     void registerUserDefinedCallback(const EnergyPlusData &state, const std::function<void(void *)> &f, const std::string &programNameInInputFile);
 
@@ -100,7 +98,6 @@ namespace PluginManagement {
         void shutdown() const;
 
         // methods
-        static void reportPythonError(EnergyPlusData &state);
         bool run(EnergyPlusData &state, EMSManager::EMSCallFrom iCallingPoint) const; // calls main() on this plugin instance
 
         // plugin calling point hooks
@@ -173,7 +170,6 @@ namespace PluginManagement {
         ~PluginManager();
 
         static int numActiveCallbacks(const EnergyPlusData &state);
-        static void addToPythonPath(EnergyPlusData &state, const fs::path &includePath, bool userDefinedPath);
         static void setupOutputVariables(EnergyPlusData &state);
 
         int maxGlobalVariableIndex = -1;
