+++ src/EnergyPlus/PythonEngine.cc
@@ -82,265 +82,6 @@ namespace Python {
 
 #if LINK_WITH_PYTHON
 
-    // RAII helper to convert a std::filesystem::path to a wchar_t* that can be passed to Python C API functions, and ensure proper cleanup of the
-    // wchar_t* if it was allocated
-    struct PyWcharPath
-    {
-        PyWcharPath(const PyWcharPath &) = delete;
-        PyWcharPath &operator=(const PyWcharPath &) = delete;
-
-        explicit PyWcharPath(const fs::path &p)
-        {
-            if constexpr (std::is_same_v<fs::path::value_type, wchar_t>) {
-                wstr_ = p.generic_wstring();
-                ptr_ = wstr_.data();
-            } else {
-                ptr_ = Py_DecodeLocale(p.generic_string().c_str(), nullptr);
-            }
-        }
-
-        ~PyWcharPath()
-        {
-            if constexpr (!std::is_same_v<fs::path::value_type, wchar_t>) {
-                PyMem_RawFree(ptr_);
-            }
-        }
-
-        const wchar_t *get() const
-        {
-            return ptr_;
-        }
-        operator const wchar_t *() const
-        {
-            return ptr_;
-        }
-
-    private:
-        wchar_t *ptr_ = nullptr;
-        std::wstring wstr_; // only populated on Windows
-    };
-
-    void reportPythonError([[maybe_unused]] EnergyPlusData &state)
-    {
-        PyObject *exc_type = nullptr;
-        PyObject *exc_value = nullptr;
-        PyObject *exc_tb = nullptr;
-        PyErr_Fetch(&exc_type, &exc_value, &exc_tb);
-        // Normalizing the exception is needed. Without it, our custom EnergyPlusException go through just fine
-        // but any ctypes built-in exception for eg will have wrong types
-        PyErr_NormalizeException(&exc_type, &exc_value, &exc_tb);
-        PyObject *str_exc_value = PyObject_Repr(exc_value); // Now a unicode object
-        PyObject *pyStr2 = PyUnicode_AsEncodedString(str_exc_value, "utf-8", "Error ~");
-        Py_DECREF(str_exc_value);
-        char *strExcValue = PyBytes_AsString(pyStr2); // NOLINT(hicpp-signed-bitwise)
-        Py_DECREF(pyStr2);
-        EnergyPlus::ShowContinueError(state, "Python error description follows: ");
-        EnergyPlus::ShowContinueError(state, strExcValue);
-
-        // See if we can get a full traceback.
-        // Calls into python, and does the same as capturing the exception in `e`
-        // then `print(traceback.format_exception(e.type, e.value, e.tb))`
-        PyObject *pModuleName = PyUnicode_DecodeFSDefault("traceback");
-        PyObject *pyth_module = PyImport_Import(pModuleName);
-        Py_DECREF(pModuleName);
-
-        if (pyth_module == nullptr) {
-            EnergyPlus::ShowContinueError(state, "Cannot find 'traceback' module in reportPythonError(), this is weird");
-            return;
-        }
-
-        PyObject *pyth_func = PyObject_GetAttrString(pyth_module, "format_exception");
-        Py_DECREF(pyth_module); // PyImport_Import returns a new reference, decrement it
-
-        if ((pyth_func != nullptr) || (PyCallable_Check(pyth_func) != 0)) {
-
-            PyObject *pyth_val = PyObject_CallFunction(pyth_func, "OOO", exc_type, exc_value, exc_tb);
-
-            // traceback.format_exception returns a list, so iterate on that
-            if ((pyth_val == nullptr) || !PyList_Check(pyth_val)) { // NOLINT(hicpp-signed-bitwise)
-                EnergyPlus::ShowContinueError(state, "In reportPythonError(), traceback.format_exception did not return a list.");
-                return;
-            }
-
-            Py_ssize_t numVals = PyList_Size(pyth_val);
-            if (numVals == 0) {
-                EnergyPlus::ShowContinueError(state, "No traceback available");
-                return;
-            }
-
-            EnergyPlus::ShowContinueError(state, "Python traceback follows: ");
-
-            EnergyPlus::ShowContinueError(state, "```");
-
-            for (Py_ssize_t itemNum = 0; itemNum < numVals; itemNum++) {
-                PyObject *item = PyList_GetItem(pyth_val, itemNum);
-                if (PyUnicode_Check(item)) { // NOLINT(hicpp-signed-bitwise) -- something inside Python code causes warning
-                    std::string traceback_line = PyUnicode_AsUTF8(item);
-                    if (!traceback_line.empty() && traceback_line[traceback_line.length() - 1] == '\n') {
-                        traceback_line.erase(traceback_line.length() - 1);
-                    }
-                    EnergyPlus::ShowContinueError(state, std::format(" >>> {}", traceback_line));
-                }
-                // PyList_GetItem returns a borrowed reference, do not decrement
-            }
-
-            EnergyPlus::ShowContinueError(state, "```");
-
-            // PyList_Size returns a borrowed reference, do not decrement
-            Py_DECREF(pyth_val); // PyObject_CallFunction returns new reference, decrement
-        }
-        Py_DECREF(pyth_func); // PyObject_GetAttrString returns a new reference, decrement it
-    }
-
-    void addToPythonPath(EnergyPlusData &state, const fs::path &includePath, bool userDefinedPath)
-    {
-        if (includePath.empty()) {
-            return;
-        }
-
-        // We use generic_string / generic_wstring here, which will always use a forward slash as directory separator even on windows
-        // This doesn't handle the (very strange, IMHO) case were on unix you have backlashes (which are VALID filenames on Unix!)
-        // Could use FileSystem::makeNativePath first to convert the backslashes to forward slashes on Unix
-        PyObject *unicodeIncludePath = nullptr;
-        if constexpr (std::is_same_v<typename fs::path::value_type, wchar_t>) {
-            const std::wstring ws = includePath.generic_wstring();
-            unicodeIncludePath = PyUnicode_FromWideChar(ws.c_str(), static_cast<Py_ssize_t>(ws.size())); // New reference
-        } else {
-            const std::string s = includePath.generic_string();
-            unicodeIncludePath = PyUnicode_FromString(s.c_str()); // New reference
-        }
-        if (unicodeIncludePath == nullptr) {
-            EnergyPlus::ShowFatalError(state, std::format("ERROR converting the path \"{:g}\" for addition to the sys.path in Python", includePath));
-        }
-
-        PyObject *sysPath = PySys_GetObject("path"); // Borrowed reference
-        int const ret = PyList_Insert(sysPath, 0, unicodeIncludePath);
-        Py_DECREF(unicodeIncludePath);
-
-        if (ret != 0) {
-            if (PyErr_Occurred() != nullptr) {
-                reportPythonError(state);
-            }
-            EnergyPlus::ShowFatalError(state, std::format("ERROR adding \"{:g}\" to the sys.path in Python", includePath));
-        }
-
-        if (userDefinedPath) {
-            EnergyPlus::ShowMessage(state, std::format("Successfully added path \"{:g}\" to the sys.path in Python", includePath));
-        }
-
-        // PyRun_SimpleString)("print(' EPS : ' + str(sys.path))");
-    }
-
-    void initPython(EnergyPlusData &state, fs::path const &programDir)
-    {
-        PyStatus status;
-
-        // first pre-config Python so that it can speak UTF-8
-        PyPreConfig preConfig;
-        // This is the other related line that caused Decent CI to start having trouble.  I'm putting it back to
-        // PyPreConfig_InitPythonConfig, even though I think it should be isolated.  Will deal with this after IO freeze.
-        PyPreConfig_InitPythonConfig(&preConfig);
-#    if DEBUG_PYTHON_CONFIG
-        EnergyPlus::print("PyPreConfig initialized:\n{}\n", preConfig);
-#    endif
-        // PyPreConfig_InitIsolatedConfig(&preConfig);
-        // PyPreConfig_InitIsolatedConfig sets configure_locale=0 which likely caused Decent CI failures
-        // https://github.com/python/cpython/blob/v3.12.2/Python/preconfig.c#L310-L345
-        preConfig.utf8_mode = 1;
-        // disable use_environment so VIRTUAL_ENV/PYTHONPATH don't leak the user's venv into EnergyPlus's embedded Python
-        // preConfig.use_environment = 0;
-#    if DEBUG_PYTHON_CONFIG
-        EnergyPlus::print("Final PyPreConfig:\n{}\n", preConfig);
-#    endif
-        status = Py_PreInitialize(&preConfig);
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not pre-initialize Python to speak UTF-8... {}", status));
-        }
-
-        PyConfig config;
-        PyConfig_InitIsolatedConfig(&config);
-
-#    if DEBUG_PYTHON_CONFIG
-        EnergyPlus::print("Isolated config initialized:\n{}\n", config);
-#    endif
-        config.isolated = 1;
-
-        PyWcharPath wcharProgramPath(FileSystem::getAbsolutePath(FileSystem::getProgramPath()));
-        status = PyConfig_SetString(&config, &config.executable, wcharProgramPath);
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not initialize executable on PyConfig... {}", status));
-        }
-
-        status = PyConfig_SetString(&config, &config.program_name, L"energyplus");
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not initialize program_name on PyConfig... {}", status));
-        }
-
-        // Ensure site.py doesn't run, this picks up your virtualenv, even with isolated config
-        // config.site_import = 0;
-
-        status = PyConfig_Read(&config);
-        if (PyStatus_Exception(status) != 0) {
-            ShowFatalError(state, std::format("Could not read back the PyConfig... {}", status));
-        }
-
-        fs::path const pathToPythonPackages = programDir / "python_lib";
-        {
-            PyWcharPath wcharPath(pathToPythonPackages);
-
-            status = PyConfig_SetString(&config, &config.home, wcharPath);
-            if (PyStatus_Exception(status) != 0) {
-                ShowFatalError(state, std::format("Could not set home to {:g} on PyConfig... {}", pathToPythonPackages, status));
-            }
-
-            status = PyConfig_SetString(&config, &config.base_prefix, wcharPath);
-            if (PyStatus_Exception(status) != 0) {
-                ShowFatalError(state, std::format("Could not set base_prefix to {:g} on PyConfig... {}", pathToPythonPackages, status));
-            }
-
-            config.module_search_paths_set = 1;
-            status = PyWideStringList_Append(&config.module_search_paths, wcharPath);
-            if (PyStatus_Exception(status) != 0) {
-                ShowFatalError(state, std::format("Could not add {:g} to module_search_paths on PyConfig... {}", pathToPythonPackages, status));
-            }
-        }
-
-        {
-            // we also need to set an extra import path to find some dynamic library loading stuff, again make it relative to the binary
-            fs::path pathToPythonLibDynload = pathToPythonPackages / "lib-dynload";
-            PyWcharPath wcharPath(pathToPythonLibDynload);
-            status = PyWideStringList_Append(&config.module_search_paths, wcharPath);
-            if (PyStatus_Exception(status) != 0) {
-                ShowFatalError(
-                    state, std::format("Could not add {:g} to module_search_paths on PyConfig... {}", pathToPythonLibDynload, status));
-            }
-        }
-
-        {
-            // we'll always want to add the program executable directory to PATH so that Python can find the installed pyenergyplus package
-            PyWcharPath wcharPath(programDir);
-            status = PyWideStringList_Append(&config.module_search_paths, wcharPath);
-            if (PyStatus_Exception(status) != 0) {
-                ShowFatalError(state, std::format("Could not add {:g} to module_search_paths on PyConfig... {}", programDir, status));
-            }
-        }
-
-#    if DEBUG_PYTHON_CONFIG
-        EnergyPlus::print("Final PyConfig:\n{}\n", config);
-#    endif
-        // This was Py_InitializeFromConfig(&config), but was giving a seg fault when running inside
-        // another Python instance, for example as part of an API run.  Per the example here:
-        // https://docs.python.org/3/c-api/init_config.html#preinitialize-python-with-pypreconfig
-        // It looks like we don't need to initialize from config again, it should be all set up with
-        // the init calls above, so just initialize and move on.
-        // UPDATE: This worked happily for me on Linux, and also when I build locally on Windows, but not on Decent CI
-        // I suspect a difference in behavior for Python versions.  I'm going to temporarily revert this back to initialize
-        // with config and get IO freeze going, then get back to solving it.
-        // Py_Initialize();
-        Py_InitializeFromConfig(&config);
-        PyConfig_Clear(&config);
-    }
-
     PythonEngine::PythonEngine(EnergyPlusData &state) : eplusRunningViaPythonAPI(state.dataPluginManager->eplusRunningViaPythonAPI)
     {
         // we'll need the program directory for a few things so get it once here at the top and sanitize it
@@ -351,12 +92,7 @@ namespace Python {
             programDir = FileSystem::getParentDirectoryPath(FileSystem::getAbsolutePath(FileSystem::getProgramPath()));
         }
 
-        initPython(state, programDir);
-
-        // now for additional paths:
-        // we will then optionally add the current working directory to allow Python to find scripts in the current directory
-        // we will then optionally add the directory of the running IDF to allow Python to find scripts kept next to the IDF
-        // we will then optionally add any additional paths the user specifies on the search paths object
+        EnergyPlus::PythonHelpers::initPython(state, programDir);
 
         PyObject *m = PyImport_AddModule("__main__");
         if (m == nullptr) {
